fix: Mood & Spot id가 Long으로 변경되면서 미처 변경하지 못한 연관된 datatype 수정

+++ src/main/java/com/poomy/mainserver/user/dto/req/RegisterUserSpotsReqDto.java
@@ -9,6 +9,6 @@ import java.util.List;
 public class RegisterUserSpotsReqDto {
 
     @Size(min = 1, max = 2, message = "사용자 핫 플레이스는 1개 이상 2개 이하여야 합니다.")
-    List<Integer> spotIds;
+    List<Long> spotIds;
 
 }
